--- assets/scss/theme.scss
@@ -1,1 +1,0 @@
-$orange: orange;
