+++ assets/style/global.scss
@@ -1,0 +1,57 @@
+@import "https://fonts.googleapis.com/icon?family=Material+Icons";
+@import "https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@48,400,1,0";
+
+@font-face {
+  font-family: Hind Madurai Light;
+  src: url("./assets/font/HindMadurai-Light.ttf") format("truetype");
+}
+
+@font-face {
+  font-family: Hind Madurai Medium;
+  src: url("./assets/font/HindMadurai-Medium.ttf") format("truetype");
+}
+
+@font-face {
+  font-family: Hind Madurai Regular;
+  src: url("./assets/font/HindMadurai-Regular.ttf") format("truetype");
+}
+
+@font-face {
+  font-family: Hind Madurai SemiBold;
+  src: url("./assets/font/HindMadurai-SemiBold.ttf") format("truetype");
+}
+
+@font-face {
+  font-family: Hind Madurai Bold;
+  src: url("./assets/font/HindMadurai-Bold.ttf") format("truetype");
+}
+
+::-webkit-scrollbar {
+  width: 0px;
+}
+
+.showWhenMobile {
+  display: none !important;
+}
+
+.showWhenDesktop {
+  display: block !important;
+}
+
+@media only screen and (max-width: 600px) {
+  .showWhenMobile {
+    display: block !important;
+  }
+
+  .showWhenDesktop {
+    display: none !important;
+  }
+}
+
+body {
+  background: #d8ecf4;
+  background-image: url("@/assets/images/background.png");
+  background-size: cover;
+  background-repeat: no-repeat;
+  font-family: Hind Madurai Regular;
+}
