+++ nuxt.config.js
@@ -18,7 +18,7 @@ export default {
 
   // Global CSS: https://go.nuxtjs.dev/config-css
   css: [
-    '~assets/scss/theme'
+    '~assets/style/global'
   ],
 
   // Plugins to run before rendering page: https://go.nuxtjs.dev/config-plugins
