+++ pages/index.vue
@@ -1,5 +1,5 @@
 <script>
-import Header from "../src/components/Header";
+import Header from "@/src/components/Header";
 export default {
   components: { Header },
   name: "IndexPage",
@@ -14,9 +14,6 @@ export default {
 <style scoped>
 .wrapper {
   height: 100vh;
-  background-image: url("../assets/images/background.png");
-  background-size: cover;
-  background-repeat: no-repeat;
   display: flex;
   justify-content: center;
   align-items: center;
