+++ src/components/Common/Tag/index.vue
